+++ index.d.ts
@@ -60,5 +60,5 @@ type RequestMeta = {
     };
 };
 type RequestPacket = Record<string, any> & RequestMeta;
-type Service = 'assess' | 'author' | 'authoraide' | 'items' | 'reports' | 'questions' | 'data';
+type Service = 'assess' | 'author' | 'authoraide' | 'events' | 'items' | 'reports' | 'questions' | 'data';
 type Action = 'get' | 'set' | 'update';